with_suffix() shouldn't prepend the curdir if file has no dir
Added test for with_suffix() on file with no dir, and fixed code so that
it works as documented, i.e. the current directory is not prepended if
the original file has no directory.

Added with_extension() as alias to with_suffix().

Added documentation for with_extension().

Improved documentation for extension() and with_suffix.

Moved tests related to basefile() outside the test loop for different
suffix styles.

Changed the test loop for different suffix styles to use an array
instead of a hash for its data, since the keys were not used anyway.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch stem
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   lib/Path/Class/File.pm
#	modified:   t/08-stem.t
#